refactor: move legacy donation forms into v2 domain
