+++ src/DonationForms/V2/ListTable/Columns/ShortcodeColumn.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Give\DonationForms\ListTable\Columns;
+namespace Give\DonationForms\V2\ListTable\Columns;
 
-use Give\DonationForms\Models\DonationForm;
+use Give\DonationForms\V2\Models\DonationForm;
 use Give\Framework\ListTable\ModelColumn;
 
 /**
@@ -12,10 +12,8 @@ use Give\Framework\ListTable\ModelColumn;
  *
  * @extends ModelColumn<DonationForm>
  */
-class StatusColumn extends ModelColumn
+class ShortcodeColumn extends ModelColumn
 {
-    protected $sortColumn = 'status';
-
     /**
      * @since 2.24.0
      *
@@ -23,7 +21,7 @@ class StatusColumn extends ModelColumn
      */
     public static function getId(): string
     {
-        return 'status';
+        return 'shortcode';
     }
 
     /**
@@ -33,7 +31,7 @@ class StatusColumn extends ModelColumn
      */
     public function getLabel(): string
     {
-        return __('Status', 'give');
+        return __('Shortcode', 'give');
     }
 
     /**
@@ -46,9 +44,8 @@ class StatusColumn extends ModelColumn
     public function getCellValue($model): string
     {
         return sprintf(
-            '<div class="statusBadge statusBadge--%1$s"><p>%2$s</p></div>',
-            $model->status->getValue(),
-            $model->status->label()
+            '<div class="shortcode">[give_form id="%d"]</div>',
+            $model->id
         );
     }
 }
