+++ src/ServiceProviders/RestAPI.php
@@ -18,9 +18,9 @@ use Give\API\Endpoints\Reports\TotalDonors;
 use Give\API\Endpoints\Reports\TotalIncome;
 use Give\API\Endpoints\Reports\TotalRefunds;
 use Give\API\RestRoute;
-use Give\DonationForms\Endpoints\FormActions;
-use Give\DonationForms\Endpoints\ListDonationForms;
-use Give\DonationForms\Endpoints\SwitchDonationFormView;
+use Give\DonationForms\V2\Endpoints\FormActions;
+use Give\DonationForms\V2\Endpoints\ListDonationForms;
+use Give\DonationForms\V2\Endpoints\SwitchDonationFormView;
 use Give\Donations\Endpoints\DonationActions;
 use Give\Donations\Endpoints\ListDonations;
 use Give\Donations\Endpoints\SwitchDonationView;
