+++ tests/Unit/DonationForms/Endpoints/TestListDonationForms.php
@@ -3,8 +3,8 @@
 namespace Give\Tests\Unit\DonationForms\Endpoints;
 
 use Exception;
-use Give\DonationForms\Endpoints\ListDonationForms;
-use Give\DonationForms\ListTable\DonationFormsListTable;
+use Give\DonationForms\V2\Endpoints\ListDonationForms;
+use Give\DonationForms\V2\ListTable\DonationFormsListTable;
 use Give\Tests\TestCase;
 use Give\Tests\TestTraits\RefreshDatabase;
 use Give\Tests\Unit\DonationForms\TestTraits\LegacyDonationFormAdapter;
