+++ tests/Unit/DonationForms/TestTraits/LegacyDonationFormAdapter.php
@@ -2,9 +2,9 @@
 
 namespace Give\Tests\Unit\DonationForms\TestTraits;
 
-use Give\DonationForms\Models\DonationForm;
-use Give\DonationForms\Properties\DonationFormLevel;
-use Give\DonationForms\ValueObjects\DonationFormStatus;
+use Give\DonationForms\V2\Models\DonationForm;
+use Give\DonationForms\V2\Properties\DonationFormLevel;
+use Give\DonationForms\V2\ValueObjects\DonationFormStatus;
 use Give\Framework\Support\Facades\DateTime\Temporal;
 use Give\Framework\Support\ValueObjects\Money;
 use Give_Donate_Form;
